chore: renamed plugin to another_flutter_recorder
- includes fixes to handle IOS crash when AVAudioRecorder destinationURL is incorrect/contains errors
- Fix to IOS and Android build
- Updated readme file
