+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'flutter_audio_recorder'
+rootProject.name = 'another_audio_recorder'
