+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.zeno.flutter_audio_recorder">
+  package="com.loonix.another_audio_recorder">
 </manifest>
