+++ android/src/main/java/com/loonix/another_audio_recorder/RequestPermissionHandler.java
@@ -1,4 +1,4 @@
-package com.zeno.flutter_audio_recorder;
+package com.loonix.another_audio_recorder;
 
 import android.Manifest;
 import android.app.Activity;
