+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]},{"name":"flutter_audio_recorder","path":"/Users/danielcarneiro/DEV/loonix/flutter_audio_recorder/","dependencies":[]},{"name":"path_provider_ios","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_ios-2.0.7/","dependencies":[]}],"android":[{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]},{"name":"flutter_audio_recorder","path":"/Users/danielcarneiro/DEV/loonix/flutter_audio_recorder/","dependencies":[]},{"name":"path_provider_android","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_android-2.0.11/","dependencies":[]}],"macos":[{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]},{"name":"path_provider_macos","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_macos-2.0.5/","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_linux-2.1.5/","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_windows-2.0.5/","dependencies":[]}],"web":[{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]}]},"dependencyGraph":[{"name":"audioplayers","dependencies":["path_provider"]},{"name":"flutter_audio_recorder","dependencies":[]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_ios","path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_ios","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2022-02-15 10:14:30.715775","version":"2.10.1"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"another_audio_recorder","path":"/Users/danielcarneiro/DEV/loonix/another_audio_recorder/","dependencies":[]},{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]},{"name":"path_provider_ios","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_ios-2.0.7/","dependencies":[]}],"android":[{"name":"another_audio_recorder","path":"/Users/danielcarneiro/DEV/loonix/another_audio_recorder/","dependencies":[]},{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]},{"name":"path_provider_android","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_android-2.0.11/","dependencies":[]}],"macos":[{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]},{"name":"path_provider_macos","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_macos-2.0.5/","dependencies":[]}],"linux":[{"name":"path_provider_linux","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_linux-2.1.5/","dependencies":[]}],"windows":[{"name":"path_provider_windows","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/path_provider_windows-2.0.5/","dependencies":[]}],"web":[{"name":"audioplayers","path":"/Users/danielcarneiro/.pub-cache/hosted/pub.dartlang.org/audioplayers-0.20.1/","dependencies":[]}]},"dependencyGraph":[{"name":"another_audio_recorder","dependencies":[]},{"name":"audioplayers","dependencies":["path_provider"]},{"name":"path_provider","dependencies":["path_provider_android","path_provider_ios","path_provider_linux","path_provider_macos","path_provider_windows"]},{"name":"path_provider_android","dependencies":[]},{"name":"path_provider_ios","dependencies":[]},{"name":"path_provider_linux","dependencies":[]},{"name":"path_provider_macos","dependencies":[]},{"name":"path_provider_windows","dependencies":[]}],"date_created":"2022-02-15 12:35:07.007103","version":"2.10.1"}
