+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,13 +1,14 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
 export "FLUTTER_ROOT=/Users/danielcarneiro/flutter"
-export "FLUTTER_APPLICATION_PATH=/Users/danielcarneiro/DEV/loonix/flutter_audio_recorder/example"
+export "FLUTTER_APPLICATION_PATH=/Users/danielcarneiro/DEV/loonix/another_audio_recorder/example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
-export "FLUTTER_TARGET=lib/main.dart"
+export "FLUTTER_TARGET=/Users/danielcarneiro/DEV/loonix/another_audio_recorder/example/lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "FLUTTER_BUILD_NAME=1.0.0"
-export "FLUTTER_BUILD_NUMBER=1"
+export "FLUTTER_BUILD_NUMBER=0000000001"
+export "DART_DEFINES=RkxVVFRFUl9XRUJfQVVUT19ERVRFQ1Q9dHJ1ZQ=="
 export "DART_OBFUSCATION=false"
-export "TRACK_WIDGET_CREATION=false"
+export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"
-export "PACKAGE_CONFIG=.packages"
+export "PACKAGE_CONFIG=/Users/danielcarneiro/DEV/loonix/another_audio_recorder/example/.dart_tool/package_config.json"
