+++ example/pubspec.yaml
@@ -1,6 +1,8 @@
-name: flutter_audio_recorder_example
-description: Demonstrates how to use the flutter_audio_recorder plugin.
+name: another_audio_recorder_example
+description: Demonstrates how to use the another_audio_recorder plugin.
 publish_to: 'none'
+
+version: 1.0.0+0000000001
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -14,7 +16,7 @@ dependencies:
 
 
 dev_dependencies:
-  flutter_audio_recorder:
+  another_audio_recorder:
     path: ../
   flutter_test:
     sdk: flutter
