+++ ios/Classes/AnotherAudioRecorderPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface AnotherAudioRecorderPlugin : NSObject<FlutterPlugin>
+@end
