--- ios/Classes/FlutterAudioRecorderPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterAudioRecorderPlugin : NSObject<FlutterPlugin>
-@end
