--- ios/Classes/FlutterAudioRecorderPlugin.m
@@ -1,8 +1,0 @@
-#import "FlutterAudioRecorderPlugin.h"
-#import <flutter_audio_recorder/flutter_audio_recorder-Swift.h>
-
-@implementation FlutterAudioRecorderPlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftFlutterAudioRecorderPlugin registerWithRegistrar:registrar];
-}
-@end
