+++ ios/another_audio_recorder.podspec
@@ -2,7 +2,7 @@
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html
 #
 Pod::Spec.new do |s|
-  s.name             = 'flutter_audio_recorder'
+  s.name             = 'another_audio_recorder'
   s.version          = '0.5.4'
   s.summary          = 'A new flutter plugin project.'
   s.description      = <<-DESC
