+++ lib/another_audio_recorder.dart
@@ -6,9 +6,9 @@ import 'package:flutter/services.dart';
 import 'package:path/path.dart' as p;
 
 /// Audio Recorder Plugin
-class FlutterAudioRecorder {
+class AnotherAudioRecorder {
   static const MethodChannel _channel =
-      const MethodChannel('flutter_audio_recorder');
+      const MethodChannel('another_audio_recorder');
   static const String DEFAULT_EXTENSION = '.m4a';
   static LocalFileSystem fs = LocalFileSystem();
 
@@ -21,7 +21,7 @@ class FlutterAudioRecorder {
   Future? get initialized => _initRecorder;
   Recording? get recording => _recording;
 
-  FlutterAudioRecorder(String path,
+  AnotherAudioRecorder(String path,
       {AudioFormat? audioFormat, int sampleRate = 16000}) {
     _initRecorder = _init(path, audioFormat!, sampleRate);
   }
