+++ test/another_audio_recorder_test.dart
@@ -1,9 +1,8 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_audio_recorder/flutter_audio_recorder.dart';
 
 void main() {
-  const MethodChannel channel = MethodChannel('flutter_audio_recorder');
+  const MethodChannel channel = MethodChannel('another_audio_recorder');
 
   setUp(() {
     channel.setMockMethodCallHandler((MethodCall methodCall) async {
